% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-bubble.R
\name{bubble}
\alias{bubble}
\title{Plot Bubble Map}
\usage{
bubble(data, shapename, choice = NULL, color.variable = NULL,
  size.variable = NULL, lgd.color.title = NULL,
  lgd.size.title = NULL, color = c("grey", "red"), size = 6,
  fill = "#e5e5e5", border.color = "black", border.size = 0.5,
  text.show = TRUE, min.segment.length = 0.5, force = 0.5,
  family = "sans", fontface = "plain", text.alpha = 1,
  text.size = 5, text.color = "black", basemap = TRUE)
}
\arguments{
\item{data}{data}

\item{shapename}{shape name variable}

\item{choice}{default is NULL}

\item{color.variable}{color variable}

\item{size.variable}{size variable}

\item{lgd.color.title}{colored legend title}

\item{lgd.size.title}{sized legend title}

\item{color}{color, which be more than one, default is grey and red}

\item{size}{size for all bubble, default is 6}

\item{fill}{fill color for base map}

\item{border.color}{map border color}

\item{border.size}{map border size}

\item{text.show}{logical, whether to show text on map, TRUE is default}

\item{min.segment.length}{\code{\link[ggrepel]{geom_text_repel}}. See the documentation for those}

\item{force}{\code{\link[ggrepel]{geom_text_repel}}. See the documentation for those}

\item{family}{\code{\link[gdtools]{font_family_exists}}. See the documentation for those}

\item{fontface}{face, plain, bold, italic}

\item{text.alpha}{alpha, default is 1}

\item{text.size}{size for text, default is 5}

\item{text.color}{text color, default is black}

\item{basemap}{logical, default is TRUE}
}
\value{
a ggplot bubble map
}
\description{
Plot Bubble Map
}
\examples{
plot(1:10)
}
