% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-heat.R
\name{heat}
\alias{heat}
\title{Plot Heat Map}
\usage{
heat(data, shapename, choice = NULL, heat.variable = NULL,
  color = c("gray", "red"), lgd.title = NULL, fill = "#e5e5e5",
  border.color = "black", border.size = 0.5, text.show = TRUE,
  min.segment.length = 0.5, force = 0.5, family = "sans",
  fontface = "plain", text.alpha = 1, text.size = 5,
  text.color = "black")
}
\arguments{
\item{data}{data}

\item{shapename}{shape name variable}

\item{choice}{default is NULL}

\item{heat.variable}{heat variable}

\item{color}{color, which be more than one, default is grey and red}

\item{lgd.title}{legend title}

\item{fill}{fill color for base map}

\item{border.color}{map border color}

\item{border.size}{map border size}

\item{text.show}{logical, whether to show text on map, TRUE is default}

\item{min.segment.length}{\code{\link[ggrepel]{geom_text_repel}}. See the documentation for those}

\item{force}{\code{\link[ggrepel]{geom_text_repel}}. See the documentation for those}

\item{family}{\code{\link[gdtools]{font_family_exists}}. See the documentation for those}

\item{fontface}{face, plain, bold, italic}

\item{text.alpha}{alpha, default is 1}

\item{text.size}{size for text, default is 5}

\item{text.color}{text color, default is black}
}
\value{
a ggplot2 heat map
}
\description{
Plot Heat Map
}
\examples{
plot(1:10)
}
