% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-map.R
\docType{data}
\name{mapdata}
\alias{mapdata}
\title{Map Data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3280 rows and 4 columns.}
\usage{
data(mapdata)
}
\description{
This map data is from gaode map, which is map in 20190523.
}
\examples{
data(mapdata)
}
\keyword{datasets}
