% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-migration.R
\name{migration}
\alias{migration}
\title{Plot Migration Map}
\usage{
migration(data, start, destination, choice = NULL,
  color.variable = NULL, size.variable = NULL, type.variable = NULL,
  lgd.color.title = NULL, lgd.size.title = NULL,
  lgd.type.title = NULL, size = 1, color = c("black", "red"),
  type = 1, fill = "#e5e5e5", border.color = "black",
  border.size = 0.5, basemap = TRUE)
}
\arguments{
\item{data}{data}

\item{start}{start position}

\item{destination}{destination}

\item{choice}{default is NULL}

\item{color.variable}{color variable}

\item{size.variable}{size variable}

\item{type.variable}{type variable}

\item{lgd.color.title}{colored legend title}

\item{lgd.size.title}{sized legend title}

\item{lgd.type.title}{typed legend title}

\item{size}{size for all line, default is 1}

\item{color}{color, which be more than one, default is black and red}

\item{type}{line type, default is 1}

\item{fill}{fill color for base map}

\item{border.color}{map border color}

\item{border.size}{map border size}

\item{basemap}{logical, default is TRUE}
}
\value{
a ggplot2 migration map
}
\description{
Plot Migration Map
}
\examples{
plot(1:10)
}
