% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nCov2019.R
\name{nCov2019}
\alias{nCov2019}
\title{New Coronavirus Data in 2019 China}
\usage{
nCov2019()
}
\value{
data of new coronavirus in 2019 china
}
\description{
New Coronavirus Data in 2019 China
}
