% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrisk.R
\name{ggrisk}
\alias{ggrisk}
\title{Risk Plot for Cox Regression}
\usage{
ggrisk(data, time, event, cutoff = "roc", lab = list(fA.ylab =
  "Risk Score", fB.ylab = "Survival Time"), legend = list(fA.title =
  "Risk Group", fB.title = "Status", fC.title = "Expression"),
  size = list(ylab = 14, fAB.points = 2, fAB.vline = 1, fAB.legendtitle =
  12, fAB.legendtext = 10), height.ratio = c(0.1, 0.1, 0.01, 0.15),
  family = "serif")
}
\arguments{
\item{data}{dataframe data}

\item{time}{variable name for following time}

\item{event}{variable name for event time}

\item{cutoff}{a number or string, which can be roc, median or cutoff}

\item{lab}{ylab for figure A and B}

\item{legend}{legend title for figure A, B and C}

\item{size}{size for y-axis labels, points, vertical line, legend title,
and legend text in figure A and B}

\item{height.ratio}{relative height}

\item{family}{family}
}
\value{
a ggplot picture
}
\description{
A ggplot style risk plot for cox regression.
}
\examples{
library(ggrisk)
ggrisk(
    data=mtcars,
    time='qsec',
    event='am'
    )
}
